slide66

+++ controllers/studentController.js
@@ -2,12 +2,23 @@ const {Student, Course, StudentCourses} = require('../models')
 
 //view all
 module.exports.viewAll = async function (req, res) {
+    if (!req.user.can('view students')){
+        res.redirect('/');
+        return
+    }
     const students = await Student.findAll();
     res.render('student/view_all', {students});
 }
 
 //profile
 module.exports.viewProfile = async function (req, res) {
+    const isAdmin = req.user.can('view students');
+    const profileBelongsToUser = req.user.can('view self') && req.user.matchesStudentId(req.params.id);
+
+    if (!isAdmin && !profileBelongsToUser) {
+        res.redirect('/')
+        return
+    }
     const student = await Student.findByPk(req.params.id, {
         include: 'courses'
     });
@@ -21,28 +32,35 @@ module.exports.viewProfile = async function (req, res) {
     res.render('student/profile', {student, availableCourses})
 }
 
-//render add
-module.exports.renderAddForm = function (req, res) {
-    const student = {
-        first_name: '',
-        last_name: '',
-        grade_level: 9,
-    }
-    res.render('student/add', {student});
-}
-
-//add
-module.exports.addStudent = async function (req, res) {
-    const student = await Student.create({
-        first_name: req.body.first_name,
-        last_name: req.body.last_name,
-        grade_level: req.body.grade_level
-    });
-    res.redirect(`/students/profile/${student.id}`);
-}
+// //render add
+// module.exports.renderAddForm = function (req, res) {
+//     const student = {
+//         first_name: '',
+//         last_name: '',
+//         grade_level: 9,
+//     }
+//     res.render('student/add', {student});
+// }
+//
+// //add
+// module.exports.addStudent = async function (req, res) {
+//     const student = await Student.create({
+//         first_name: req.body.first_name,
+//         last_name: req.body.last_name,
+//         grade_level: req.body.grade_level
+//     });
+//     res.redirect(`/students/profile/${student.id}`);
+// }
 
 //render edit
 module.exports.renderEditForm = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesStudentId(req.params.id);
+
+    if(!isAdmin && !profileBelongsToUser) {
+        res.redirect('/')
+        return
+    }
     const student = await Student.findByPk(req.params.id);
     console.log(student);
     res.render('student/edit', {student});
@@ -50,6 +68,13 @@ module.exports.renderEditForm = async function (req, res) {
 
 //update
 module.exports.updateStudent = async function (req, res) {
+    const isAdmin = req.user.can('edit student');
+    const profileBelongsToUser = req.user.can('edit self') && req.user.matchesStudentId(req.params.id);
+
+    if(!isAdmin && !profileBelongsToUser) {
+        res.redirect('/')
+        return
+    }
     const student = await Student.update({
         first_name: req.body.first_name,
         last_name: req.body.last_name,
@@ -65,6 +90,10 @@ module.exports.updateStudent = async function (req, res) {
 
 //delete
 module.exports.deleteStudent = async function (req, res) {
+    if (!user.can('delete student')){
+        res.redirect('/');
+        return
+    }
     await Student.destroy({
         where: {
             id: req.params.id
